Optionally skip absolute fee check (#491)
Having a more lax policy is useful for pay-to-open, because the sender may not retry payments.

The previous implementation in #487 was a bit too lax though, it created an edge case when amount was less than the pay-the-open fee. The whole payment was eaten by the LSP in that case.