add exception handler in io coroutines
It is good practice to execute blocking io-related calls inside the `Dispatchers.IO` context, however errors should be properly handled to not make the `Dispatchers.IO` context crash, which is a fatal error.

According to [this guide](https://medium.com/@hakim.fakher/kotlin-coroutines-basics-part-2-4087ce15ff28), we should define our own exception handler and execute the code in a subroutine of the `Dispatchers.IO` context.