Rework connection flow
We clean up the coroutine hierarchy: the `ElectrumClient` has a single
internal coroutine that is launched once the connection has been
established with the electrum server. This coroutine launches three child
coroutines and uses supervision to gracefully stop if the connection is
closed or an error is received.

Connection establishment happens in the context of the job that calls
`connect()` and doesn't throw exceptions.

We revert the introduction of a `CoroutineExceptionHandler` inside the
JVM socket code, as it shouldn't be needed anymore.