Fix iOS tests
When connecting to an invalid address in tests on iOS, the error isn't
reported by the native socket which just hangs until the test times out.

We lower the timeout on the `connect` call, which previously had the same
value as the test timeout. We thus fail the connection attempt after
1 second and are able to reconnect to a different server.

The very low RPC timeout was flaky on iOS: because the clock used depends
on the actual dispatcher, it sometimes triggered only after receiving a
valid response from the server, which made the test fail. Setting it to
0ms ensures that the is immediately cancelled with a timeout exception
regardless of the dispatcher used.