trust local mutual close for swap-in
Instead of explicitly providing txids, we can look at local channel data
to extract mutual close txids, because either the tx is unconfirmed and
the channel data will still be in `Closing` state, otherwise the tx is
confirmed and the utxo can be used immediately anyway.

This is an alternative to #516.