add methods to start/stop swap-in wallet
For the migraton use case, it is useful to be able to wait for a certain
number of UTXOs to be published before doing a swap-in. The reason is
that:
- we want to make as few on-chain operations as possible to save fees,
  so it's better to group
- migration swap-in have special characteristics that are only
  implemented for channel creation, not for splices.

The new methods `startWatchSwapInWallet`/`stopWatchSwapInWallet` are
footgunny and should only be used in the context of the migration, to be
removed later. Note that the `Peer` calls `startWatchSwapInWallet` at
initialization, so the migration process must first `stop...()`, then
later `start...()`.