Process actions synchronously
This commit contains a very subtle change. When processing an event, we
previously added the follow-up actions to the `input` channel. Since other
coroutines could concurrently add events to that `input` channel, that
means we were asynchronously processing follow-up actions and it was hard
to evaluate how independent events could affect processing.

We now synchronously process all the actions generated by a given event.
We're conceptually moving from a model where our execution tree was
executed using BFS to a DFS execution. It makes a lot more sense, as each
event is now fully processed before moving on to the next queued event.