Emit Disconnected before reconnection
We previously emitted the `Disconnected` event in a different coroutine
than the one we use to reconnect. If that coroutine was starved, this
could result in the `Disconnected` event being added to the queue after
the `Connected` event for the new connection, which results in an invalid
state.

We now emit `Disconnected` and `Connected` from the same coroutine, which
guarantees the correct ordering between those two events. It feels a bit
weird because if we don't reconnect, we won't emit the `Disconnected`
event, but this is ok, as we only really need to clean up state from the
previous connection once we reestablish.