Fix `invalid_commit_sig` issue when expecting `tx_signatures` (#536)
* Always retransmit `tx_signatures` if requested

If our peer asks us to retransmit our `tx_signatures`, we should do it even
if the transaction is already confirmed.

* Ignore previously received `commit_sig`

When expecting a retransmission of `tx_signatures`, we should ignore the
`commit_sig` they send just before if we've already received it. The right
way to check that we've already received it is to compare its signature
to our latest commitment transaction.