Exposed swap-in utxos refund delay
Utxos used for swap-in have an expiry: after a delay the user can spend
them unilaterally, which makes them dangerous to accept by the swap
provider when we get close to that expiry. We expose these limits so
that applications can tell their users to finalize swaps before reaching
the expiry, otherwise they'll have to wait before unlocking them.

Fixes #534