Don't double-spend unspendable inputs (#559)
When filtering inputs that can be used for swaps, we filter out all the
inputs that are already used in pending funding transactions. However,
we ignored previously confirmed funding transactions, assuming that our
electrum server would have already filtered them from our wallet state.

Unfortunately, we cannot assume that our electrum server is up-to-date
(especially if we connect to a different one after a restart) and behaving
correctly, so we must also exclude inputs use in the latest confirmed
funding transaction.