unlock utxos when peer is unavailable for splice
If the peer is offline when the swap-in job requests a splice-in, the
request will be ignored, but we weren't unlocking UTXOs, preventing any
further attempt.

Also add an optimistic synchronization in the swap-in job, which as a
side effect will cause us to retry swaps everytime the peer gets
reconnected (vs at every new block or liquidity policy update previously).