Send batch_size on commit_sig retransmit
If we get disconnected after sending `commit_sig`, we will retransmit
that message when reconnecting if our peer has not received it.

When we have multiple commitments, we need to retransmit one message
per commitment, and include the `batch_size` tlv.

We incorrectly stored `commit_sig` without the `batch_size` tlv in the
next remote commit field, and thus retransmitted without that tlv.