Purge pending incoming payments on disconnection (#583)
We were keeping pending incoming payments in the internal state of the
`IncomingPaymentHandler` even when disconnected. This could lead to
inconsistencies with the channel state: if an HTLC that was added to the
`IncomingPaymentHandler` was failed by another instance of the app (for
example a background task), that failure wouldn't propagate to the state
of the `IncomingPaymentHandler`. This can lead to partially fulfilled
payments, where we think we have received the complete payment, but one
of the part was actually failed so we shouldn't release the preimage.