Fix parsing error and make script hash status nullable
When a script has no history, its status as returned by Electrum is "null" which we did not parse properly.
It had no real impact on our wallet since electrum would return an empty history or unspent items list.
Checking that a status is null can be used to detect unused addresses: even if there are no UTXOs, a script
that has received transactions will have a non-null status.