Add EncodedNodeId (#601)
Adds the EncodedNodeId interface that can be either a public key or a pair channel id, direction that's more compact.
This is used as introduction node for blinded routes.
We also use EncodedNodeId for the next node to relay a message to. Which is not in the spec yet but is necessary for us since we have no way to resolve the compact node id to a public key ourselves. It requires a compatible peer that accepts this format.