Rotate taproot swap-in addresses (#584)
Addresses are generated one by one, as soon as we detect that the last generated address has been used. There is no lookahead.

Scanning 'synchronous' (not really synchronous, but we rely the least possible on the mailbox). This makes the `window` disappear because we are scanning addresses one by one.

Some cleanup by extracting wallets outside of `Peer`.

---------

Co-authored-by: sstone <fabrice.drouin@acinq.fr>