Refactor key management - fundingKey/paymentKey
We want to reduce key material manipulation to crypto related code.
The goal is for the rest of the code to only manipulate key
descriptions but never actual keys.

Pros:
* Reduces the scope of the code that stores and manipulates
  secrets;
* Opens the door to delegating secrets management to external
  components;
* Serves as an anti-corruption layer between lightning code
  and crypto library.

This first commit only focuses on the fundingKey and the paymentKey
as they seem the easiest to handle.

We introduce `sign2` as an alternative to `sign` (we do not look
for a better name as it is supposed to replace `sign` at the end of
the refactoring). The `sign2` function takes a `PrivateKeyDescriptor`
instead of a `PrivateKey`. The idea is to migrate all calls from
`sign` to `sign2`. During the migration, we keep both functions
to ensure that code compiles and tests pass.

`LocalPrivateKeyDescriptor` is put inside `LocalKeyManager` as it is
where it is needed now. It could well be the case that it's not the
smartest place but it's not clear yet if we'll keep this class so
let's not spend too much time on where to put it yet.

Regarding the tests, they build test keys directly from hex strings
so we introduce another implementation of `PrivateKeyDescriptor` in
the tests (`HardCodedPrivateKey`) to encapsulate them.