Refactor key management - swap in protocol
Focuses on the swap in protocol part which relies on extended keys.
So we introduce a new type of descriptor: `ExtendedPrivateKeyDescriptor`.
We also introduce its implementation: `DerivedExtendedPrivateKeyDescriptor`.

At some point, we have a chicken and egg issue with key descriptors
describing how to create a key by deriving a parent: what about the
parent of all keys? We introduce `RootExtendedPrivateKeyDescriptor`
to represent the root master key.

There are two methods that are duplicated in
`DerivedExtendedPrivateKeyDescriptor` and in
`RootExtendedPrivateKeyDescriptor`:
* `publicKey()`;
* `derivePrivateKey()`.

We keep this duplication for now as we're not sure how these two classes
will end up.

Regarding the tests, some of them encode the keys as `ByteArray` so
we introduce a new constructor to `HardCodedPrivateKey` to handle
that case.