(starblocks) Add endpoint for static website and add package gradle task
Website is now in resources directory. The API delivers the static
website on "/". The API endpoints are behing "/api".

The executable expects the website files to be in the "/web" directory
but this can be customised using the "webdir" argument.

New gradle tasks have been added to build the website and the server
executable, and put them in the "build/package" directory.

For example:
=> ./gradlew :starblocks:packageMacosX64