Refactor key management - segregate instantiate() again
We remove `instantiate()` from `ExtendedPrivateKeyDescriptor` so
that only the `LocalKeyManager` can instantiate actual keys.

To ease the process, we introduce derivation capabilities to
`ExtendedPrivateKeyDescriptor`.

We change `LocalKeyManager` to only rely on this way of
deriving keys. Some part of the code still extract key material
so it may look a bit clunky with casts to be able to call the
`instantiate()` method. Let's keep it that way until we find
a better technique because we're happy with the idea of:
* not exposing `instantiate()` to the outside world ;
* relying on one single implementation of each type of derivation.