Add on-the-fly fee credit feature
We add an optional feature that lets on-the-fly funding clients accept
payments that are too small to pay the fees for an on-the-fly funding.
When that happens, the payment amount is added as "fee credit" without
performing an on-chain operation. Once enough fee credit has been
obtained, we can initiate an on-chain operation to create a channel or
a splice.

This feature makes more efficient use of on-chain transactions by
trusting that our peer will honor our fee credit in the future.