Rework `invoice_request` timeout mechanism
We may want to display a spinner in the UI while requesting the invoice
for an offer payment, to quickly report back when the recipient cannot
be contacted. This means we need to fail as soon as the timeout provided
is hit, so we cannot rely on a regular background task.

We instead create a dedicated coroutine for each payment, with the
timeout provided in the `PayOffer` request.