Generate default deterministic offer (#626)
Add the option to generate a default, deterministic offer that can be
restored based on the node's seed and the trampoline node it uses.
This offer has the minimal amount of information necessary. It doesn't
expire, which makes it suitable to save for long-term, recurring
payments.

When we include a blinded path in the offer, it doesn't make sense to
include an `offer_node_id`. When receiving an `invoice_request`, we
will signed the corresponding invoice with the blinded private key
matching the blinded path used by the sender.