Refactor `OfferManager`
This commit keeps the business logic mostly unchanged, but re-works the
error handling to ensure that we log as much details as possible to help
troubleshoot compatibility issues.

It adds support for sending `invoice_error` on some error cases to help
debug potential issues.

When the introduction node of the recipient is our trampoline node, we
remove an unnecessary hop in the onion message.