More detailed error when HTLC could not be sent
We use more detailed errors when we fail to send an HTLC, since this
may happen because the channel is offline, syncing, opening or closing.

It may also happen that we have a channel in the `Normal` state that
has enough funds, but we cannot send the HTLC because it hits a channel
limit (too many HTLCs pending, a splice is ongoing, etc). When that
happens we will retry ignoring the channel, which leads to the default
`NoAvailableChannels` error: in that case, the "real" error can be found
in the `failures` list of the `OutgoingPaymentFailure` instance.

Fixes #616