Convert partial failure into user-friendly errors
Some payment errors can be converted into friendly user errors, telling
them whether they should retry or not and what should change in their
payment attempt.

This isn't the case for all errors: some of them cannot be interpreted
easily when they come from remote nodes, but that's ok, we're doing
the best we can when the error can be interpreted.