resolve feerates as late as possible
Instead of waiting for feerates to be available before processing any
channel command, we just pass the `StateFlow` to the `ChannelContext`
and will attempt to get feerates only when needed.

If feerates are not available, the `ChannelState.process` function will
hang. Previously, the hanging would have happened in the `Peer`.