Unblock channels stuck in `Syncing`
When using channel backups, we wait for our peer to send their
`channel_reestablish` first. This can be an issue for channels that
are only present locally, and have been forgotten by our peer (for
example because they force-close while we were offline).

In that case we would stay in `Syncing` indefinitely: to avoid that,
we send our `channel_reestablish` if we haven't received theirs, which
gives them an opportunity to send back an error and forget the channel.