Back `PeerConnection` with a flow (#675)
Since #668, `connectionState` is strictly managed by the underlying TCP connection, not by the main peer loop. As a result, it is not usable in tests where we use a fake connection. In particular the function `waitForPeerReady()` will never return.

We back the private mutable `PeerConnection` by a `StateFlow` and use this instead of `connectionState`.