handle feature-less bolt11 invoices
We were not handling correctly feature-less invoices, which are
spec-compliant. We do reject invoices that don't have
`var_onion_optin`, but the check happens later. Instead, we currently
throw a `NullPointerException`.

The way we are checking requirements in the `init` block requires
lazy-init properties, otherwise the object initialization will fail
before any logic can happen.

Also, as per the spec, the feature tag must be skipped if there are no
features enabled.