Add per-contact deterministic payer keys
When paying people we added as trusted contacts, we may want to use a
deterministic payer_key which lets them know the payment came from us.
If they have added us to their own contact list, they will be able to
match the payment and know it came from us.

This is of course optional: whenever we don't want to reveal that we
are the source of a payment, we keep using a random payer_key.