Add support for official trampoline payments to blinded paths
We update the blinded path TLVs to use the official spec values.
We include blinded paths in the _outer_ onion instead of previously
including them in the _trampoline_ onion. This allows paying nodes
that support trampoline, for whom we can include a trampoline payload.
This lets the sender provide arbitrary TLVs to the recipient, even when
using trampoline.

To verify that the trampoline onion really comes from the intended
sender, the onion packet for the recipient uses an associated data with
a shared secret from the `invoice_request` in its HMAC. Without this
protection, the trampoline node could replace the trampoline onion with
one that it created.