Update trampoline payment to blinded path to match spec proposal
We update our trampoline payments to blinded paths to match the official
specification from https://github.com/lightning/bolts/pull/836.

The blinded paths and recipient features are included in the trampoline
onion, which potentially allows using multiple trampoline hops.

That was already what we were doing with experimental TLVs, so we simply
update the TLV values to match the spec values.