move payment id to WalletPayment
This allows for a unified identifier for all payments that affect the
balance.

For `IncomingPayment`, the `payment_hash` is currently abusively used as
a primary key, which is hacky in the context of swap-ins. To minimize
integration effort, the payment id simply derived from that existing
key, with no other changes.