Include local mining fees for pure liquidity purchases (#710)
When we are purchasing liquidity without any other splice operation
(splice-in, splice-out or splice-cpfp), we will have a single entry
in our DB: the liquidity purchase itself. In that case, we are paying
mining fees for the shared input and output: we must record those in
the purchase. This happens during manual liquidity purchases and
on-the-fly funding.

We cannot easily add tests because we don't implement the liquidity
seller side, which makes it hard to trigger those conditions.