Increase default invoice expiry
We increase the default expiry of Bolt 11 and Bolt 12 invoices: they
are now valid for 1 day. This ensures that if a payment takes a while
to reach us (most likely because we've been offline for a few hours),
the invoice is still valid when we come back online to receive the
payment.

We also increase our `min_final_expiry_delta` to 144 blocks and our
`fulfill_safety_before_timeout` to 12 blocks. This gives us more time
to publish our commitment transaction and HTLC-success transaction in
case our peer doesn't acknowledge a preimage we reveal, so that we can
claim the payment on-chain.