Add failing test for delayed HTLC relay
When using on-the-fly funding, we have to create and publish an on-chain
transaction after accepting `will_add_htlc`, before we can relay the
matching HTLCs. This can create issues when there isn't enough margin
in the `min_final_expiry_delta` parameter where the HTLC is rejected.

We add failing tests that demonstrate this behavior.