Fix `min_final_expiry_delta` failures for on-the-fly funding
We fix the failing tests added in the previous commit, by allowing
on-the-fly HTLCs that are below `min_final_expiry_delta` as long as
their expiry is greater than twice our `fulfill_safety_before_timeout`
parameter. This is what we want to guarantee: that we have enough time
to potentially force-close and claim HTLC-success transactions.