Remove `max_final_expiry_delta` parameter
This field was not at all doing what was documented: it doesn't make
sense to reject a payment because its `min_final_expiry_delta` is too
large, we only care about the total path's expiry delta which cannot
exceed 2016 blocks.

We were using this for the blinded path expiry, but the spec says we
must rather use the invoice expiry, roughly converted to a block height.

We also reorder `minFinalExpiryDelta` to put it closer to other related
parameters.